+++ lib/service/Analytics.dart
@@ -41,6 +41,7 @@ import 'package:rokwire_plugin/service/analytics.dart' as rokwire;
 import 'package:rokwire_plugin/model/poll.dart';
 import 'package:rokwire_plugin/model/group.dart';
 import 'package:illinois/ext/Group.dart';
+import 'package:illinois/ext/Favorite.dart';
 import 'package:illinois/service/Auth2.dart';
 
 import 'package:illinois/service/NativeCommunicator.dart';
@@ -277,6 +278,7 @@ class Analytics extends rokwire.Analytics implements NotificationsListener {
 
   // Event Attributes
   static const String   LogAttributeUrl                    = "url";
+  static const String   LogAttributeSource                 = "source";
   static const String   LogAttributeEventId                = "event_id";
   static const String   LogAttributeEventName              = "event_name";
   static const String   LogAttributeEventCategory          = "event_category";
@@ -842,15 +844,12 @@ class Analytics extends rokwire.Analytics implements NotificationsListener {
   }
 
   void logFavorite(Favorite? favorite, {bool? on, String? title}) {
-    if (on == null) {
-      on = Auth2().isFavorite(favorite);
-    }
     logEvent({
       LogEventName          : LogFavoriteEventName,
-      LogFavoriteActionName : on ? LogFavoriteOnActionName : LogFavoriteOffActionName,
+      LogFavoriteActionName : (on ?? Auth2().isFavorite(favorite)) ? LogFavoriteOnActionName : LogFavoriteOffActionName,
       LogFavoriteTypeName   : favorite?.favoriteKey,
       LogFavoriteIdName     : favorite?.favoriteId,
-      LogFavoriteTitleName  : title,
+      LogFavoriteTitleName  : title ?? favorite?.favoriteTitle,
     });
   }