+++ lib/service/Guide.dart
@@ -4,6 +4,7 @@ import 'dart:io';
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart';
+import 'package:illinois/service/Analytics.dart';
 import 'package:rokwire_plugin/model/auth2.dart';
 import 'package:rokwire_plugin/service/app_datetime.dart';
 import 'package:rokwire_plugin/service/app_livecycle.dart';
@@ -261,6 +262,14 @@ class Guide with Service implements NotificationsListener {
     return ((result != null) && (stripHtmlTags == true)) ? StringUtils.stripHtmlTags(result) : result;
     // Bidi.stripHtmlIfNeeded(result);
   }
+
+  Map<String, dynamic>? entryAnalyticsAttributes(Map<String, dynamic>? entry) => (entry != null) ? {
+    Analytics.LogAttributeGuideId : entryId(entry),
+    Analytics.LogAttributeGuideTitle : JsonUtils.stringValue(Guide().entryTitle(entry, stripHtmlTags: true)),
+    Analytics.LogAttributeGuide : JsonUtils.stringValue(Guide().entryValue(entry, 'guide')),
+    Analytics.LogAttributeGuideCategory :  JsonUtils.stringValue(Guide().entryValue(entry, 'category')),
+    Analytics.LogAttributeGuideSection :  JsonUtils.stringValue(Guide().entryValue(entry, 'section')),
+  } : null;
 
   bool isEntryReminder(Map<String, dynamic>? entry) {
     return JsonUtils.stringValue(entryValue(entry, 'content_type')) == campusReminderContentType;